6877 Componentize Data Tool Geography Info box
